Made Styling Scalable for Application
Big big big commit. I moved a styles.js file and a common component for
a button to a Common directory. This makes it easier to find and use
both the button and styling. Plus, when we want to add more styles, we
just add another entry into the styles.js line.